user-service/requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.0.3
Flask-JWT-Extended==4.5.2
psycopg2-binary==2.9.6
redis==4.5.5
python-dotenv==1.0.0
